move receipt creation to separate classes
- add methods to order interface and move receipt creation logic to
separate classes
- create interfaces for Order, Line, Bike to increase testability
- add test coverage for new classes and new order interface

Testing Order, Line, and Bike behavior was difficult (or impossible) to
do with all the dependencies on concrete classes. Adding interfaces
allows for the creating of mock objects so the tests can focus
exclusively on testing the behavior of only the specific class under
test, and introduces flexibility for future development.

The Order class was handling too much. Receipt creation is a separate
concern and should be handled by separate classes which are focused on
that single responsibility.